Method for aligning similar clusters over time
This is the code that will rearrange clusters according to the cosine similarity between the current row and the prior row. The method is "greedy" and will look for the most similar available cluster. In this manner, the cluster number with the lowest number should generally have less volatile changes, while the cluster with the highest number may be the least stable of the clusters.

Using the cluster numbers (prior and current; although they are removed in this version) may permit linking back to the original records so as to evaluate the contents of the cluster based either on attributes used for the constructing the cluster or other attributes that may give additional insight into the contents of the cluster.